테스트 파일 시도, psql로는 넘어오는 것 확인하여 ORM 방식 버림. psycopg2를 사용한 뷰로 수정 환경변수, 기존 db 부분(라우터, settings.py) 주석 처리

+++ restaurants/models.py
@@ -1,23 +1,23 @@
 from django.db import models
 
 class Restaurant(models.Model):
-    id = models.IntegerField(primary_key=True)
-    name = models.CharField(max_length=255)  # 음식점 이름
-    status = models.CharField(max_length=50)  # 상태
-    address_zip_code = models.CharField(max_length=50, null=True, blank=True)  # 우편번호
-    road_zip_code = models.CharField(max_length=25, null=True, blank=True)  # 도로명 우편번호
-    road_full_address = models.CharField(max_length=500, null=True, blank=True)  # 도로명 전체 주소
-    road_address = models.CharField(max_length=500, null=True, blank=True)  # 도로명 주소
-    x = models.FloatField(null=True, blank=True)  # x좌표
-    y = models.FloatField(null=True, blank=True)  # y좌표
-    phone_number = models.CharField(max_length=50, null=True, blank=True)  # 전화번호
-    category_1 = models.CharField(max_length=50, null=True, blank=True)  # 카테고리 1
-    category_2 = models.CharField(max_length=50, null=True, blank=True)  # 카테고리 2
-    district_name = models.CharField(max_length=50, null=True, blank=True)  # 구/군 이름
-    attribute_1 = models.CharField(max_length=5, null=True, blank=True)  # 속성 1
-    attribute_2 = models.CharField(max_length=5, null=True, blank=True)  # 속성 2
-    attribute_3 = models.CharField(max_length=5, null=True, blank=True)  # 속성 3
-    attribute_4 = models.CharField(max_length=5, null=True, blank=True)  # 속성 4
+    id = models.BigAutoField(primary_key=True)
+    name = models.CharField(max_length=255)
+    status = models.CharField(max_length=50)
+    address_zip_code = models.CharField(max_length=50, null=True, blank=True)
+    road_zip_code = models.CharField(max_length=25, null=True, blank=True)
+    road_full_address = models.CharField(max_length=500, null=True, blank=True)
+    road_address = models.CharField(max_length=500, null=True, blank=True)
+    x = models.FloatField(null=True, blank=True)
+    y = models.FloatField(null=True, blank=True)
+    phone_number = models.CharField(max_length=50, null=True, blank=True)
+    category_1 = models.CharField(max_length=50, null=True, blank=True)
+    category_2 = models.CharField(max_length=50, null=True, blank=True)
+    district_name = models.CharField(max_length=50, null=True, blank=True)
+    attribute_1 = models.CharField(max_length=5, null=True, blank=True)
+    attribute_2 = models.CharField(max_length=5, null=True, blank=True)
+    attribute_3 = models.CharField(max_length=5, null=True, blank=True)
+    attribute_4 = models.CharField(max_length=5, null=True, blank=True)
 
     def __str__(self):
         return self.name
